requirements.txt
python-nmap
scapy
prettytable 